% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categorical_field_response_values.R
\name{get_categorical_field_response_values}
\alias{get_categorical_field_response_values}
\title{Generate categorical field response values}
\usage{
get_categorical_field_response_values(categorical_field_responses)
}
\arguments{
\item{categorical_field_responses}{A long dataframe of categorical
field response values and weights.}
}
\value{
A long dataframe of categorical field response values with one
row for each value set.
}
\description{
Provide a set of response values for each categorical field in
`categorical_field_responses`
}
\examples{
\dontrun{
get_categorical_field_response_values(categorical_field_responses)
}
}
\keyword{Internal}
