% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categorical_field_responses.R
\name{get_categorical_field_responses}
\alias{get_categorical_field_responses}
\title{Get every categorical field response from a REDCap data dictionary}
\usage{
get_categorical_field_responses(metadata)
}
\arguments{
\item{metadata}{A REDCap data dictionary}
}
\value{
a dataframe with these columns
\describe{
  \item{field_name}{First item}
  \item{field_type}{Second item}
  \item{response_code}{Second item}
  \item{response_label}{Second item}
  \item{field_group}{Second item}
  \item{weight}{a set of uniform weights across the responses of each field}
}
}
\description{
Given a REDCap data dictionary, enumerate every response value for every categorical field in that data dictionary
}
\examples{
\dontrun{
categorical_field_responses <-
  get_categorical_field_responses(metadata_to_populate)
}
}
