% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_long_categorical_field_response_values.R
\name{get_long_categorical_field_response_values}
\alias{get_long_categorical_field_response_values}
\title{generate categorical field response values}
\usage{
get_long_categorical_field_response_values(long_categorical_field_responses)
}
\arguments{
\item{long_categorical_field_responses}{a long data set of categorical
field response values and weights.}
}
\value{
a tall dataframe of categorical field response values with one
row for each value set.
}
\description{
Provide a set of response values for each categorical field in
`long_categorical_field_responses`
}
\examples{
\dontrun{
get_long_categorical_field_response_values(long_categorical_field_responses)
}
}
