% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_long_notes_fields.R
\name{get_long_notes_fields}
\alias{get_long_notes_fields}
\title{Get long notes fields from a REDCap data dictionary}
\usage{
get_long_notes_fields(metadata)
}
\arguments{
\item{metadata}{A REDCap data dictionary}
}
\value{
A dataframe with the following columns:
\describe{
  \item{field_name}{The name of the field.}
  \item{form_name}{The form name in which the field is located.}
  \item{field_type}{The type of the field (will always be "notes").}
  \item{weight}{A uniform weight, set to 100.}
}
}
\description{
Given a REDCap data dictionary, this function returns the fields (rows) that have a field type of notes.
}
\examples{
\dontrun{
notes_fields <- get_long_notes_fields(metadata_to_populate)
}
}
