% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_long_text_field_values.R
\name{get_long_text_field_values}
\alias{get_long_text_field_values}
\title{generate text field values}
\usage{
get_long_text_field_values(long_text_fields)
}
\arguments{
\item{long_text_fields}{a long data set of text
fields, their parameters, and weights.}
}
\value{
a long dataframe of text field values with one row for each value set.
}
\description{
Provide a set of values for each field in the output of
`get_long_text_fields`
}
\examples{
\dontrun{
get_long_text_field_values(long_text_fields)
}
}
