% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_notes_field_values.R
\name{get_notes_field_values}
\alias{get_notes_field_values}
\title{Generate notes field content}
\usage{
get_notes_field_values(notes_fields)
}
\arguments{
\item{notes_fields}{A long dataframe of notes fields from a REDCap data dictionary
(typically output from [get_notes_fields()])}
}
\value{
A tall dataframe of notes field content with one row per field
}
\description{
Generate lorem ipsum text content for each notes field in `notes_fields`
}
\examples{
\dontrun{
notes_responses <- get_notes_field_values(notes_fields)
}
}
\keyword{Internal}
