% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_notes_fields.R
\name{get_notes_fields}
\alias{get_notes_fields}
\title{Get notes fields from a REDCap data dictionary and annotate them with weights}
\usage{
get_notes_fields(metadata)
}
\arguments{
\item{metadata}{A REDCap data dictionary}
}
\value{
A long dataframe with the following columns:
\describe{
  \item{field_name}{The name of the field.}
  \item{form_name}{The form name in which the field is located.}
  \item{field_type}{The type of the field (this will always be _notes_).}
  \item{weight}{A uniform weight, set to 100.}
}
}
\description{
Given a REDCap data dictionary, this function returns the rows that have a field type of _notes_.
}
\examples{
\dontrun{
notes_fields <- get_notes_fields(metadata_to_populate)
}
}
\keyword{Internal}
