% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_rectangle_of_values.R
\name{get_one_rectangle_of_values}
\alias{get_one_rectangle_of_values}
\title{Generate a rectangle of data for one record}
\usage{
get_one_rectangle_of_values(
  one_record_id = 1,
  record_id_name,
  forms_to_fill,
  long_categorical_field_responses,
  long_notes_fields_responses
)
}
\arguments{
\item{one_record_id}{a single record_id}

\item{record_id_name}{the column name the record_id should be returned in}

\item{forms_to_fill}{the forms to fill for this rectangle}

\item{long_categorical_field_responses}{the output of `get_long_categorical_field_responses()`}

\item{long_notes_fields_responses}{the output of `get_long_notes_fields()`}
}
\value{
a rectangle of data with appropriate REDCap identifiers ready to write to REDCap
}
\description{
Generate a rectangle of data for one record
}
\examples{
\dontrun{
get_one_rectangle_of_values(1, record_id_name, forms_to_fill)
}
}
