% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_values.R
\name{get_project_values}
\alias{get_project_values}
\title{Get Project Values}
\usage{
get_project_values(
  redcap_uri,
  token,
  events = NULL,
  number_of_records_to_populate = 5
)
}
\arguments{
\item{redcap_uri}{REDCap API URI}

\item{token}{REDCap project API token}

\item{events}{An optional vector of event names to fill (defaults to all events)}

\item{number_of_records_to_populate}{An optional number of records to generate (default 5)}
}
\value{
A list of data frames (one per event, or just one for classic
projects). Each list element contains a dataframe of synthetic data
for each of the forms on that event. Each dataframe is ready to load
into REDCap.
}
\description{
Generate a dataset of synthetic data for a REDCap project using the project
design retrieved via the REDCap API.
}
