% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slider_field_values.R
\name{get_slider_field_values}
\alias{get_slider_field_values}
\title{Generate slider field values}
\usage{
get_slider_field_values(slider_values)
}
\arguments{
\item{slider_values}{A long data set of slider values and weights.}
}
\value{
A long dataframe of slider values with one row for each field
}
\description{
Generate values for each slider field in `slider_values`
}
\examples{
\dontrun{
get_slider_field_values(slider_values)
}
}
\keyword{Internal}
