% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text_fields.R
\name{get_text_fields}
\alias{get_text_fields}
\title{Get every text field from a REDCap data dictionary}
\usage{
get_text_fields(metadata)
}
\arguments{
\item{metadata}{A REDCap data dictionary}
}
\value{
a dataframe with these columns
\describe{
  \item{field_name}{REDCap field name}
  \item{form_name}{REDCap form name}
  \item{field_type}{REDCap field type}
  \item{text_validation_type}{REDCap text validation type}
  \item{text_validation_min}{REDCap text validation min}
  \item{text_validation_max}{REDCap text validation max}
  \item{tvt}{text validation type function name}
  \item{weight}{a default weight for the field}
  \item{mean}{mean of data to be generated derived from the min and max}
  \item{sd}{standard deviation of data to be generated derived from the min and max}
}
}
\description{
Given a REDCap data dictionary, enumerate every text field in that data dictionary
along with validation types, min, and max, and append a text_validation_type (tvt)
function name, weight, mean, and standard deviation as needed.
}
\examples{
\dontrun{
text_fields <-
  get_text_fields(metadata_to_populate)
}
}
\keyword{Internal}
