% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapfiller-package.R
\docType{package}
\name{redcapfiller-package}
\alias{redcapfiller}
\alias{redcapfiller-package}
\title{redcapfiller: Fill a REDCap Project with Generated Data}
\description{
Fill a REDCap project with generated data based on the project design and a minimal set of inputs.
}
\author{
\strong{Maintainer}: Philip Chase \email{pbc@ufl.edu} (\href{https://orcid.org/0000-0002-5318-9420}{ORCID})

Authors:
\itemize{
  \item Sai Pavan Kamma \email{saipavankamma@ufl.edu} (\href{https://orcid.org/0009-0004-4619-0409}{ORCID})
  \item Laurence James-Woodley \email{lawjames1@ufl.edu} (\href{https://orcid.org/0000-0002-6418-2742}{ORCID})
}

Other contributors:
\itemize{
  \item Taryn Stoffs \email{tls@ufl.edu} (\href{https://orcid.org/0000-0002-0830-8179}{ORCID}) [contributor]
  \item Christopher Barnes \email{cpb@ufl.edu} (\href{https://orcid.org/0000-0001-7114-1992}{ORCID}) [contributor]
}

}
\keyword{internal}
