% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_values.R
\name{get_project_values}
\alias{get_project_values}
\title{Get Project Values}
\usage{
get_project_values(
  redcap_uri,
  token,
  events = NULL,
  number_of_records_to_populate = 5
)
}
\arguments{
\item{redcap_uri}{REDCap API URI}

\item{token}{REDCap project API token}

\item{events}{(Optional) Vector of event names to fill}

\item{number_of_records_to_populate}{Number of records to generate (default 5)}
}
\value{
List of data frames (one per event, or just one for classic projects)
}
\description{
Get Project Values
}
