% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slider_fields.R
\name{get_slider_fields}
\alias{get_slider_fields}
\title{Get slider field ranges from a REDCap data dictionary}
\usage{
get_slider_fields(metadata)
}
\arguments{
\item{metadata}{A REDCap data dictionary}
}
\value{
a dataframe with these columns
\describe{
  \item{field_name}{REDCap field name}
  \item{form_name}{REDCap form name}
  \item{field_type}{REDCap field type}
  \item{select_choices_or_calculations}{RedCap select_choices_or_calculations field}
  \item{text_validation_min}{REDCap text validation min}
  \item{text_validation_max}{REDCap text validation max}
  \item{mean}{mean of data to be generated}
  \item{sd}{standard deviation of data to be generated}

}
}
\description{
Given a REDCap data dictionary, enumerate range of values for every slider field in that data dictionary
}
\examples{
\dontrun{
slider_fields <-
  get_slider_fields(metadata_to_populate)
}
}
