% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapfiller-package.R
\docType{package}
\name{redcapfiller-package}
\alias{redcapfiller}
\alias{redcapfiller-package}
\title{redcapfiller: Fill a REDCap Project with synthetic data}
\description{
Fill a REDCap project with synthetic data based on the project design and a minimal set of inputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ctsit.github.io/redcapfiller/}
  \item \url{https://github.com/ctsit/redcapfiller/}
  \item Report bugs at \url{https://github.com/ctsit/redcapfiller/issues/}
}

}
\author{
\strong{Maintainer}: Philip Chase \email{pbc@ufl.edu} (\href{https://orcid.org/0000-0002-5318-9420}{ORCID})

Authors:
\itemize{
  \item Sai Pavan Kamma \email{saipavankamma@ufl.edu} (\href{https://orcid.org/0009-0004-4619-0409}{ORCID})
  \item Laurence James-Woodley \email{lawjames1@ufl.edu} (\href{https://orcid.org/0000-0002-6418-2742}{ORCID})
}

Other contributors:
\itemize{
  \item Taryn Stoffs \email{tls@ufl.edu} (\href{https://orcid.org/0000-0002-0830-8179}{ORCID}) [contributor]
  \item Natalie Williams (\href{https://orcid.org/0009-0004-2983-1542}{ORCID}) [contributor]
  \item William Beasley \email{whb4@ou.edu} (\href{https://orcid.org/0000-0002-5613-5006}{ORCID}) [contributor]
  \item Aidan Wilson \email{aidan.wilson@intersect.org.au} (\href{https://orcid.org/0000-0001-9858-5470}{ORCID}) [contributor]
  \item Christopher Battiston \email{chris.battiston@wchospital.ca} (\href{https://orcid.org/0000-0003-2179-3735}{ORCID}) [contributor]
  \item Christopher Barnes \email{cpb@ufl.edu} (\href{https://orcid.org/0000-0001-7114-1992}{ORCID}) [contributor]
}

}
\keyword{internal}
